Refactor media chat component: restructure ChatMessageList and related components
- Moved ChatMessageList to a new directory for better organization.
- Created new components: DateDivider, GroupedMediaRow, MessageBubble, MessageHeader, ReplyPreview, SideActions, SingleMessageRow, and buildRenderItems utility.
- Enhanced rendering logic for messages, including grouping media messages and handling replies.
- Improved optimistic UI updates for message reactions.
- Updated imports and component structure for clarity and maintainability.
